Improve UX and UI for the HomePage.
