+++ src/App.js
@@ -9,13 +9,14 @@ import CarDetail from './components/pages/car/CarDetail';
 import CarDelete from './components/pages/car/CarDelete';
 import RentalNew from './components/pages/rentals/RentalNew';
 import SignUp from './components/pages/sessions/SignUp';
-import HomePage from './components/pages/HomePage';
+import HomePage from './components/pages/Home/HomePage';
 import ProtectedRoute from './components/ProtectedRoute';
 import Navigation from './components/shared/Navigation';
 
 const App = () => {
   const active = useSelector((state) => state.credentials.active);
 
+  /* eslint-disable */
   return (
     <BrowserRouter>
       <header>
@@ -26,67 +27,67 @@ const App = () => {
         <Routes>
           <Route
             path="/login"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={!active} redirectTo="/">
                 <Login />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/signup"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={!active} redirectTo="/">
                 <SignUp />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={active} redirectTo="/login">
                 <HomePage />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/newcar"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={active} redirectTo="/login">
                 <NewCar />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/deletecar"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={active} redirectTo="/login">
                 <CarDelete />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/car/:id"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={active} redirectTo="/login">
                 <CarDetail />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/rentals"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={active} redirectTo="/login">
                 <Rentals />
               </ProtectedRoute>
-            )}
+            }
           />
           <Route
             path="/rental/new"
-            element={(
+            element={
               <ProtectedRoute isAuthenticated={active} redirectTo="/login">
                 <RentalNew />
               </ProtectedRoute>
-            )}
+            }
           />
         </Routes>
       </main>
